requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]
psycopg2-binary==2.9.9
python-jose[cryptography]
passlib[bcrypt]
alembic
python-multipart